refactor: move Recaptcha to be a rule
