+++ app/Http/Controllers/Auth/RegisteredUserController.php
@@ -6,9 +6,9 @@ namespace App\Http\Controllers\Auth;
 
 use App\Jobs\DownloadUserAvatar;
 use App\Models\User;
+use App\Rules\Recaptcha;
 use App\Rules\Username;
 use App\Rules\ValidTimezone;
-use App\Services\Recaptcha;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -32,7 +32,7 @@ final readonly class RegisteredUserController
      *
      * @throws \Illuminate\Validation\ValidationException
      */
-    public function store(Request $request, Recaptcha $recaptcha): RedirectResponse
+    public function store(Request $request): RedirectResponse
     {
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
@@ -40,16 +40,8 @@ final readonly class RegisteredUserController
             'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:'.User::class],
             'timezone' => ['required', 'string', 'max:255', new ValidTimezone],
             'password' => ['required', 'confirmed', Rules\Password::defaults()],
-            'g-recaptcha-response' => app()->environment('production') ? ['required', 'string'] : [],
+            'g-recaptcha-response' => ['required', new Recaptcha($request->ip())],
         ]);
-
-        $ipAddress = $request->ip();
-
-        assert(is_string($ipAddress));
-
-        if (app()->environment('production') && ! $recaptcha->verify($ipAddress, $request->string('g-recaptcha-response')->value())) {
-            return back()->withErrors(['g-recaptcha-response' => 'The recaptcha response was invalid.']); // @codeCoverageIgnore
-        }
 
         $user = User::create([
             'name' => $request->name,
