+++ bootstrap/providers.php
@@ -5,5 +5,4 @@ declare(strict_types=1);
 return [
     App\Providers\AppServiceProvider::class,
     App\Providers\PulseServiceProvider::class,
-    App\Providers\RecaptchaServiceProvider::class,
 ];
